+++ examples/geospatial/geoJSONexample.js
@@ -21,7 +21,7 @@ const data = [
 ];
 
 
-mongoose.connect('mongodb://localhost/locations', function(err) {
+mongoose.connect('mongodb://127.0.0.1/locations', function(err) {
   if (err) {
     throw err;
   }