+++ examples/statics/statics.js
@@ -12,7 +12,7 @@ const Person = mongoose.model('Person');
 run().catch(console.error);
 
 async function run() {
-  await mongoose.connect('mongodb://localhost/persons');
+  await mongoose.connect('mongodb://127.0.0.1/persons');
   const bill = await Person.create({
     name: 'bill',
     age: 25,