+++ test/model.test.js
@@ -20,7 +20,6 @@ const EmbeddedDocument = mongoose.Types.Subdocument;
 const MongooseError = mongoose.Error;
 
 describe('Model', function() {
-
   let db;
   let Comments;
   let BlogPost;
@@ -80,8 +79,8 @@ describe('Model', function() {
     db = start();
   });
 
-  after(function() {
-    db.close();
+  after(async function() {
+    await db.close();
   });
 
   afterEach(() => util.clearTestData(db));
@@ -3719,31 +3718,26 @@ describe('Model', function() {
     });
 
     it('with positional notation on path not existing in schema (gh-1048)', function(done) {
-      const db = start();
-
       const M = db.model('Test', Schema({ name: 'string' }));
-      db.on('open', function() {
-        const o = {
-          name: 'gh-1048',
-          _id: new mongoose.Types.ObjectId(),
-          databases: {
-            0: { keys: 100, expires: 0 },
-            15: { keys: 1, expires: 0 }
-          }
-        };
+      const o = {
+        name: 'gh-1048',
+        _id: new mongoose.Types.ObjectId(),
+        databases: {
+          0: { keys: 100, expires: 0 },
+          15: { keys: 1, expires: 0 }
+        }
+      };
 
-        M.collection.insertOne(o, function(err) {
+      M.updateOne({ _id: o._id }, o, { upsert: true, strict: false }, function(err) {
+        assert.ifError(err);
+        M.findById(o._id, function(err, doc) {
           assert.ifError(err);
-          M.findById(o._id, function(err, doc) {
-            db.close();
-            assert.ifError(err);
-            assert.ok(doc);
-            assert.ok(doc._doc.databases);
-            assert.ok(doc._doc.databases['0']);
-            assert.ok(doc._doc.databases['15']);
-            assert.equal(doc.databases, undefined);
-            done();
-          });
+          assert.ok(doc);
+          assert.ok(doc._doc.databases);
+          assert.ok(doc._doc.databases['0']);
+          assert.ok(doc._doc.databases['15']);
+          assert.equal(doc.databases, undefined);
+          done();
         });
       });
     });
@@ -4329,40 +4323,49 @@ describe('Model', function() {
     });
   });
 
-  it('save max bson size error with buffering (gh-3906)', async function() {
-    this.timeout(10000);
-    const db = start({ noErrorListener: true });
-    const Test = db.model('Test', { name: Object });
+  describe('max bson size error', function() {
+    let db;
 
-    const test = new Test({
-      name: {
-        data: (new Array(16 * 1024 * 1024)).join('x')
+    afterEach(async() => {
+      if (db != null) {
+        await db.close();
+        db = null;
       }
     });
 
-    const error = await test.save().then(() => null, err => err);
+    it('save max bson size error with buffering (gh-3906)', async function() {
+      this.timeout(10000);
+      db = start({ noErrorListener: true });
+      const Test = db.model('Test', { name: Object });
 
-    assert.ok(error);
-    assert.equal(error.name, 'MongoServerError');
-    await db.close();
-  });
+      const test = new Test({
+        name: {
+          data: (new Array(16 * 1024 * 1024)).join('x')
+        }
+      });
 
-  it('reports max bson size error in save (gh-3906)', async function() {
-    this.timeout(10000);
-    const db = await start({ noErrorListener: true });
-    const Test = db.model('Test', { name: Object });
+      const error = await test.save().then(() => null, err => err);
 
-    const test = new Test({
-      name: {
-        data: (new Array(16 * 1024 * 1024)).join('x')
-      }
+      assert.ok(error);
+      assert.equal(error.name, 'MongoServerError');
     });
 
-    const error = await test.save().then(() => null, err => err);
+    it('reports max bson size error in save (gh-3906)', async function() {
+      this.timeout(10000);
+      db = await start({ noErrorListener: true });
+      const Test = db.model('Test', { name: Object });
 
-    assert.ok(error);
-    assert.equal(error.name, 'MongoServerError');
-    await db.close();
+      const test = new Test({
+        name: {
+          data: (new Array(16 * 1024 * 1024)).join('x')
+        }
+      });
+
+      const error = await test.save().then(() => null, err => err);
+
+      assert.ok(error);
+      assert.equal(error.name, 'MongoServerError');
+    });
   });
 
   describe('insertMany()', function() {
@@ -4789,6 +4792,67 @@ describe('Model', function() {
         assert.equal(error, null);
         done();
       });
+    });
+
+    it('insertMany() validation error with ordered false and rawResult for checking which documents failed (gh-12791)', async function() {
+      const schema = new Schema({
+        name: { type: String, required: true },
+        year: { type: Number, required: true }
+      });
+      const Movie = db.model('Movie', schema);
+
+      const id1 = new mongoose.Types.ObjectId();
+      const id2 = new mongoose.Types.ObjectId();
+      const id3 = new mongoose.Types.ObjectId();
+      const arr = [
+        { _id: id1, foo: 'The Phantom Menace', year: 1999 },
+        { _id: id2, name: 'The Force Awakens', bar: 2015 },
+        { _id: id3, name: 'The Empire Strikes Back', year: 1980 }
+      ];
+      const opts = { ordered: false, rawResult: true };
+      const res = await Movie.insertMany(arr, opts);
+      // {
+      //   acknowledged: true,
+      //   insertedCount: 1,
+      //   insertedIds: { '0': new ObjectId("63b34b062cfe38622738e510") },
+      //   mongoose: { validationErrors: [ [Error], [Error] ] }
+      // }
+      assert.equal(res.insertedCount, 1);
+      assert.equal(res.insertedIds[0].toHexString(), id3.toHexString());
+      assert.equal(res.mongoose.validationErrors.length, 2);
+      assert.ok(res.mongoose.validationErrors[0].errors['name']);
+      assert.ok(!res.mongoose.validationErrors[0].errors['year']);
+      assert.ok(res.mongoose.validationErrors[1].errors['year']);
+      assert.ok(!res.mongoose.validationErrors[1].errors['name']);
+    });
+
+    it('insertMany() validation error with ordered false and rawResult for mixed write and validation error (gh-12791)', async function() {
+      const schema = new Schema({
+        name: { type: String, required: true, unique: true },
+        year: { type: Number, required: true }
+      });
+      const Movie = db.model('Movie', schema);
+      await Movie.init();
+
+      const arr = [
+        { foo: 'The Phantom Menace', year: 1999 },
+        { name: 'The Force Awakens', bar: 2015 },
+        { name: 'The Empire Strikes Back', year: 1980 },
+        { name: 'The Empire Strikes Back', year: 1980 }
+      ];
+      const opts = { ordered: false, rawResult: true };
+      const err = await Movie.insertMany(arr, opts).then(() => null, err => err);
+
+      assert.ok(err);
+      assert.equal(err.insertedDocs.length, 1);
+      assert.equal(err.insertedDocs[0].name, 'The Empire Strikes Back');
+      assert.equal(err.writeErrors.length, 1);
+      assert.equal(err.writeErrors[0].index, 3);
+      assert.equal(err.mongoose.validationErrors.length, 2);
+      assert.ok(err.mongoose.validationErrors[0].errors['name']);
+      assert.ok(!err.mongoose.validationErrors[0].errors['year']);
+      assert.ok(err.mongoose.validationErrors[1].errors['year']);
+      assert.ok(!err.mongoose.validationErrors[1].errors['name']);
     });
 
     it('insertMany() populate option (gh-9720)', async function() {
@@ -5272,60 +5336,6 @@ describe('Model', function() {
           assert.equal(changeData.operationType, 'insert');
           assert.equal(changeData.fullDocument.name, 'Child');
         });
-
-        it('watch() before connecting (gh-5964)', async function() {
-          const db = start();
-
-          const MyModel = db.model('Test5964', new Schema({ name: String }));
-
-          // Synchronous, before connection happens
-          const changeStream = MyModel.watch();
-          const changed = new global.Promise(resolve => {
-            changeStream.once('change', data => resolve(data));
-          });
-
-          await db;
-          await MyModel.create({ name: 'Ned Stark' });
-
-          const changeData = await changed;
-          assert.equal(changeData.operationType, 'insert');
-          assert.equal(changeData.fullDocument.name, 'Ned Stark');
-        });
-
-        it('watch() close() prevents buffered watch op from running (gh-7022)', async function() {
-          const db = start();
-          const MyModel = db.model('Test', new Schema({}));
-          const changeStream = MyModel.watch();
-          const ready = new global.Promise(resolve => {
-            changeStream.once('data', () => {
-              resolve(true);
-            });
-            setTimeout(resolve, 500, false);
-          });
-
-          changeStream.close();
-          await db;
-          const readyCalled = await ready;
-          assert.strictEqual(readyCalled, false);
-        });
-
-        it('watch() close() closes the stream (gh-7022)', async function() {
-          const db = await start();
-          const MyModel = db.model('Test', new Schema({ name: String }));
-
-          await MyModel.init();
-
-          const changeStream = MyModel.watch();
-          const closed = new global.Promise(resolve => {
-            changeStream.once('close', () => resolve(true));
-          });
-
-          await MyModel.create({ name: 'Hodor' });
-
-          changeStream.close();
-          const closedData = await closed;
-          assert.strictEqual(closedData, true);
-        });
       });
 
       describe('sessions (gh-6362)', function() {
@@ -5359,7 +5369,6 @@ describe('Model', function() {
         });
 
         it('startSession() before connecting', async function() {
-
           const db = start();
 
           const MyModel = db.model('Test', new Schema({ name: String }));
@@ -5375,6 +5384,7 @@ describe('Model', function() {
 
           session.endSession();
 
+          await db.close();
         });
 
         it('sets session when pulling a document from db', async function() {
@@ -6767,7 +6777,8 @@ describe('Model', function() {
         );
 
       });
-      xit('creates indexes only when they do not exist on the mongodb server (gh-12250)', async() => {
+
+      it('creates indexes only when they do not exist on the mongodb server (gh-12250)', async() => {
         const userSchema = new Schema({
           name: { type: String }
         }, { autoIndex: false });
@@ -6781,15 +6792,14 @@ describe('Model', function() {
         const createIndexSpy = sinon.spy(User.collection, 'createIndex');
         const listIndexesSpy = sinon.spy(User.collection, 'listIndexes');
 
-        // Act
         await User.syncIndexes();
+
         assert.equal(createIndexSpy.callCount, 1);
-        assert.equal(listIndexesSpy.callCount, 2);
+        assert.equal(listIndexesSpy.callCount, 1);
 
         await User.syncIndexes();
 
-        // Assert
-        assert.equal(listIndexesSpy.callCount, 4);
+        assert.equal(listIndexesSpy.callCount, 2);
         assert.equal(createIndexSpy.callCount, 1);
       });
     });
@@ -6916,10 +6926,12 @@ describe('Model', function() {
           granularity: 'hours'
         },
         autoCreate: false,
+        autoIndex: false,
         expireAfterSeconds: 86400
       });
 
-      const Test = db.model('Test', schema);
+      const Test = db.model('Test', schema, 'Test');
+      await Test.init();
 
       await Test.collection.drop().catch(() => {});
       await Test.createCollection();
@@ -7454,119 +7466,6 @@ describe('Model', function() {
       const explainResult = await Model.exists({}, { explain: true });
       assert.ok(explainResult);
     });
-  });
-
-  it('Model.validate() (gh-7587)', async function() {
-    const Model = db.model('Test', new Schema({
-      name: {
-        first: {
-          type: String,
-          required: true
-        },
-        last: {
-          type: String,
-          required: true
-        }
-      },
-      age: {
-        type: Number,
-        required: true
-      },
-      comments: [{ name: { type: String, required: true } }]
-    }));
-
-
-    let err = null;
-    let obj = null;
-
-    err = await Model.validate({ age: null }, ['age']).
-      then(() => null, err => err);
-    assert.ok(err);
-    assert.deepEqual(Object.keys(err.errors), ['age']);
-
-    err = await Model.validate({ name: {} }, ['name']).
-      then(() => null, err => err);
-    assert.ok(err);
-    assert.deepEqual(Object.keys(err.errors), ['name.first', 'name.last']);
-
-    obj = { name: { first: 'foo' } };
-    err = await Model.validate(obj, ['name']).
-      then(() => null, err => err);
-    assert.ok(err);
-    assert.deepEqual(Object.keys(err.errors), ['name.last']);
-
-    obj = { comments: [{ name: 'test' }, {}] };
-    err = await Model.validate(obj, ['comments']).
-      then(() => null, err => err);
-    assert.ok(err);
-    assert.deepEqual(Object.keys(err.errors), ['comments.name']);
-
-    obj = { age: '42' };
-    await Model.validate(obj, ['age']);
-    assert.strictEqual(obj.age, 42);
-  });
-
-  it('Model.validate(...) validates paths in arrays (gh-8821)', async function() {
-    const userSchema = new Schema({
-      friends: [{ type: String, required: true, minlength: 3 }]
-    });
-
-    const User = db.model('User', userSchema);
-
-    const err = await User.validate({ friends: [null, 'A'] }).catch(err => err);
-
-    assert.ok(err.errors['friends.0']);
-    assert.ok(err.errors['friends.1']);
-
-  });
-
-  it('Model.validate() works with arrays (gh-10669)', async function() {
-    const testSchema = new Schema({
-      docs: [String]
-    });
-
-    const Test = db.model('Test', testSchema);
-
-    const test = { docs: ['6132655f2cdb9d94eaebc09b'] };
-
-    const err = await Test.validate(test);
-    assert.ifError(err);
-  });
-
-  it('Model.validate(...) uses document instance as context by default (gh-10132)', async function() {
-    const userSchema = new Schema({
-      name: {
-        type: String,
-        required: function() {
-          return this.nameRequired;
-        }
-      },
-      nameRequired: Boolean
-    });
-
-    const User = db.model('User', userSchema);
-
-    const user = new User({ name: 'test', nameRequired: false });
-    const err = await User.validate(user).catch(err => err);
-
-    assert.ifError(err);
-
-  });
-  it('Model.validate(...) uses object as context by default (gh-10346)', async() => {
-
-    const userSchema = new mongoose.Schema({
-      name: { type: String, required: true },
-      age: { type: Number, required() {return this && this.name === 'John';} }
-    });
-
-    const User = db.model('User', userSchema);
-
-    const err1 = await User.validate({ name: 'John' }).then(() => null, err => err);
-    assert.ok(err1);
-
-    const err2 = await User.validate({ name: 'Sam' }).then(() => null, err => err);
-    assert.ok(err2 === null);
-
   });
 
   it('sets correct `Document#op` with `save()` (gh-8439)', function() {